requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
Pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
pydantic==2.5.0
opencv-contrib-python==4.9.0.80
scikit-learn==1.3.2

